requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
decorator==5.2.1
docx2pdf==0.1.8
fire==0.7.1
Flask==3.0.3
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
fonttools==4.59.2
fpdf==1.7.2
greenlet==3.2.4
gunicorn==23.0.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.1
MarkupSafe==3.0.2
moviepy==2.1.2
mysql-connector-python==9.4.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pdf2docx==0.5.8
pdf2image==1.17.0
pillow==10.4.0
proglog==0.1.12
pydub==0.25.1
PyMuPDF==1.26.4
python-docx==1.2.0
python-dotenv==1.1.1
pywin32==311
SQLAlchemy==2.0.43
termcolor==3.1.0
tqdm==4.67.1
typing_extensions==4.15.0
Werkzeug==3.0.4
